Fix handling of unknown minor faction names
The eddb minor faction API download does not include power factions. Now
it has better handling of unknown minor faction names by automatically
adding an entry to the minor faction table.  It has internal id number
as well as eddb id number.  If eddb id is null then it is not found in
the eddb download.